+++ django_afip/exceptions.py
@@ -11,3 +11,11 @@ class AfipException(Exception):
             response.Errors.Err[0].Code,
             parsers.parse_string(response.Errors.Err[0].Msg),
         ))
+
+
+class CertificateExpiredException(Exception):
+    """
+    Raised when an authentication was attempted with an expired
+    certificate.
+    """
+    pass