+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ releases, in reverse chronological order.
 2.4.0
 -----
 
+* Raise ``CertificateExpiredException`` when attempting to authenticate using
+  an expired certificate.
 * The field ``ReceiptEntry.amount`` has been renamed to ``quantity``.
 * Add a links to documentation on where to obtain the AFIP WS certificates.
 * Introduce this changelog.